Add type=label to mask and attributes in MVoxMeshGenerator
This should allow segmentations to be used in addition to labelmaps
and filter out other volumes that should not be used as a mask
or attributes labelmap. The segmentation should be automatically
exported as a labelmap for input to the mvox command.
Note: Ensure that the color table of the segmentation is correct.